Add files via upload
Choices game
This project is about a person in an island and he faces 2 doors in front of him and he has to make a choice of
which door he will enter. If he enters the red door, he will find 3 boxes, a white box, red box and a black one. If he
chooses the white box, he will unfortunately find this box filled with snakes. If he chooses the green one, he/she
will get a treasure. If he chooses the black box, he will find lots of spiders facing him. If he does not  make any
of these choices he will not find any thing and the phrase "invalid choice" would be printed